requirements.txt
flask==2.0.1
langchain==0.0.350
python-dotenv==1.0.0
pypdf==3.17.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.36.2
accelerate==0.25.0
bitsandbytes==0.41.1 